Split context modules
