+++ app/routes/__main/agreement/index.tsx
@@ -3,7 +3,7 @@ import React, { useState, ChangeEvent, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 /* Internal dependencies */
-import { SettingsContext } from '~/context/settingsContextProvider';
+import { SettingsContext } from '~/context/settingsContext/settingsContextProvider';
 import { PrimaryButton } from '~/components/common/Button';
 import { Checkbox } from '~/components/common/Checkbox';
 import { PartialView } from '~/components/PartialView';
